refactor: modal 디렉터리에 index.js 생성
