+++ src/view/web/layers/modal/index.js
@@ -1,0 +1,2 @@
+export { default as createErrorAlertModal } from "./errorAlert/createErrorAlertModal.js";
+export { default as createPrizeResultModal } from "./prizeResult/createPrizeResultModal.js";
