+++ src/view/web/layers/modal/prizeResult/prizeResultModal.css
